contratos y eliminado funcional

+++ src/components/cargas/creadores/DeleteCarga.tsx
@@ -1,5 +1,5 @@
 import { Box, Button, Typography } from "@mui/material";
-import { useContext } from "react";
+import { useContext, useState } from "react";
 import { ContextoCargas } from "../containers/ContainerTajetasCargas";
 import { ContextoGeneral } from "../../Contexto";
 import ClearSharpIcon from "@mui/icons-material/ClearSharp";
@@ -8,12 +8,24 @@ export default function DeleteCarga(props: any) {
     const { handleCloseDialog } = props;
     const { backendURL, theme } = useContext(ContextoGeneral);
     const { cargaSeleccionada } = useContext(ContextoCargas);
+    const [errorMessage, setErrorMessage] = useState<string | null>(null);
 
     const handleNoClick = () => {
         handleCloseDialog();
     };
 
     const borrarCarga = () => {
+        if (!cargaSeleccionada) return;
+
+        // Verificar si existen cupos asignados con valores mayores a 0
+        const tieneCupos = cargaSeleccionada.cupos?.some((cupo: { cupos: number; }) => cupo.cupos > 0);
+
+        if (tieneCupos) {
+            setErrorMessage("No se puede eliminar la carga. Primero elimine los cupos relacionados.");
+            return;
+        }
+
+        // Si no hay cupos asociados, proceder con la eliminación
         fetch(`${backendURL}/cargas/${cargaSeleccionada.id}`, {
             method: "DELETE",
             headers: {
@@ -22,10 +34,10 @@ export default function DeleteCarga(props: any) {
             },
         })
             .then((response) => response.json())
-
             .catch((error) => {
                 console.error("Error al borrar la carga", error);
             });
+
         handleCloseDialog();
     };
 
@@ -46,7 +58,7 @@ export default function DeleteCarga(props: any) {
                 }}
             >
                 <Typography variant="h6" color="textPrimary" align="center">
-                    ¿Está seguro de que quiere eliminar la carga?
+                    {errorMessage ? errorMessage : "¿Está seguro de que quiere eliminar la carga?"}
                 </Typography>
                 <ClearSharpIcon
                     onClick={handleCloseDialog}
@@ -68,18 +80,20 @@ export default function DeleteCarga(props: any) {
                         marginTop: 2,
                     }}
                 >
-                    <Button
-                        variant="text"
-                        onClick={borrarCarga}
-                        sx={{
-                            color: "#d68384",
-                            width: "100%",
-                            maxWidth: "100px",
-                            fontWeight: "bold",
-                        }}
-                    >
-                        Sí
-                    </Button>
+                    {!errorMessage && (
+                        <Button
+                            variant="text"
+                            onClick={borrarCarga}
+                            sx={{
+                                color: "#d68384",
+                                width: "100%",
+                                maxWidth: "100px",
+                                fontWeight: "bold",
+                            }}
+                        >
+                            Sí
+                        </Button>
+                    )}
                     <Button
                         variant="text"
                         sx={{
@@ -90,7 +104,7 @@ export default function DeleteCarga(props: any) {
                         }}
                         onClick={handleNoClick}
                     >
-                        No
+                        {errorMessage ? "Cerrar" : "No"}
                     </Button>
                 </Box>
             </Box>
